Refactor
- Moved regex utils and hnhiring parser to utils folder
- Renamed files following camelCase convention
